+++ src/main/java/com/biovizion/prestamo911/repository/FacturaRepository.java
@@ -6,5 +6,5 @@ import org.springframework.stereotype.Repository;
 import com.biovizion.prestamo911.entities.FacturaEntity;
 
 @Repository
-public interface FacturaRepository extends JpaRepository<FacturaEntity, Integer> {
+public interface FacturaRepository extends JpaRepository<FacturaEntity, Long> {
 }
