+++ src/main/java/com/biovizion/prestamo911/service/UsuarioService.java
@@ -10,11 +10,11 @@ public interface UsuarioService {
     List<UsuarioEntity> findAll();
     UsuarioEntity save(UsuarioEntity usuario);
 
-    Optional<UsuarioEntity> findById(Integer id);
+    Optional<UsuarioEntity> findById(Long id);
     Optional<UsuarioEntity> findByEmail(String email);
-    public Optional<UsuarioEntity> get(Integer id);
+    public Optional<UsuarioEntity> get(Long id);
     public void update(UsuarioEntity usuario);
-    public void delete(Integer id);
+    public void delete(Long id);
